requirements.txt
DateTime==4.3
Flask==1.1.2
imbalanced-learn==0.7.0
imblearn==0.0
joblib==0.17.0
numpy==1.19.4
pandas==1.1.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.4
spacy==2.3.2
SQLAlchemy==1.3.20